+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Most recent releases are shown at the top. Each release shows:
 ### new:
 - support for custom metadata in vectorstore (#126)
 - basic full-text indexing (#132)
+- support for using sparse vector stores with `LLM.ask` (#136)
 
 ### changed
 - **Breaking Changes**: refactor vectorstore (#133, 1e84f46)